\name{EQUIVNONINF-internal}
\alias{pwexact}
\alias{rejmax}
\alias{rjpbnc7}
\alias{teststat}
\alias{XCRIT}
\alias{FINDSIZE}
\alias{asympt_confbs}
\alias{ex_prob_ge}
\alias{ex_prob_le}
\alias{exact_confb_l}
\alias{exact_confb_r}
\alias{ex_prob_ghe}
\alias{ex_prob_lhe}
\alias{posterior}
\alias{XCRIT2}
\alias{posterior2}
\alias{posterior3}
\alias{NPLCRIT}
\alias{FINDSIZE2}
\alias{POW_NULLALT}
\alias{ppost_singleobs}
\alias{pwexacta}
\alias{rejmaxaeq}
\alias{rjpbnc}
\title{Internal Equiv_Noninf Functions}
\description{
  Internal Equiv_Noninf functions
}
\details{
These functions are not to be called by the user
}
\keyword{ internal }

